requirements.txt
flask==3.0.0
flask-cors==4.0.0
music21==9.1.0
Pillow==10.1.0
